feat(projects): Implementa scroll infinito y skeleton loading en la vista de proyectos
