+++ package.json
@@ -18,6 +18,7 @@
     "@fortawesome/free-solid-svg-icons": "^7.0.0",
     "@fortawesome/vue-fontawesome": "^3.1.1",
     "@heroicons/vue": "^2.2.0",
+    "@supabase/supabase-js": "^2.56.0",
     "@vuepic/vue-datepicker": "^11.0.2",
     "@vueuse/core": "^13.1.0",
     "axios": "^1.9.0",
