+++ src/assets/main.css
@@ -27,6 +27,10 @@ h3,
 .h3 {
   @apply text-2xl font-semibold;
 }
+h4,
+.h4 {
+  @apply text-xl font-semibold;
+}
 /*texto nomal*/
 p {
   @apply text-base;
