+++ src/components/ui/skeletons/ProjectCardSkeleton.vue
@@ -1,0 +1,23 @@
+<template>
+  <div class="rounded-lg bg-white p-6 shadow-lg dark:bg-gray-800 dark:text-gray-200">
+    <div class="animate-pulse">
+      <div class="w-full h-48 bg-gray-300 dark:bg-gray-700 rounded-t-lg"></div>
+
+      <div class="p-4">
+        <div class="h-6 bg-gray-300 dark:bg-gray-700 rounded w-3/4 mb-2"></div>
+        <div class="h-4 bg-gray-300 dark:bg-gray-700 rounded w-full mb-2"></div>
+        <div class="h-4 bg-gray-300 dark:bg-gray-700 rounded w-5/6"></div>
+
+        <div class="mt-4 flex flex-wrap gap-2">
+          <div class="h-6 bg-gray-300 dark:bg-gray-700 rounded-full w-16"></div>
+          <div class="h-6 bg-gray-300 dark:bg-gray-700 rounded-full w-20"></div>
+          <div class="h-6 bg-gray-300 dark:bg-gray-700 rounded-full w-12"></div>
+        </div>
+      </div>
+    </div>
+  </div>
+</template>
+
+<script setup lang="ts">
+// No se necesita script aquí.
+</script>
