+++ src/router/index.ts
@@ -1,24 +1,45 @@
 import { createRouter, createWebHistory } from 'vue-router'
 import HomeView from '../views/HomeView.vue'
-import PreviewComponents from '@/views/PreviewComponents.vue'
 import { useToastStore } from '@/stores/toast'
+import BaseLayout from '@/layout/BaseLayout.vue'
 
 const router = createRouter({
   history: createWebHistory(import.meta.env.BASE_URL),
   routes: [
     {
       path: '/',
-      name: '',
-      component: PreviewComponents,
+      component: BaseLayout, // Utiliza el layout como componente principal
+      children: [
+        {
+          path: '',
+          name: 'home',
+          component: HomeView,
+        },
+        {
+          path: '/about',
+          name: 'about',
+          component: () => import('@/views/AboutView.vue'),
+        },
+        {
+          path: '/proyectos',
+          name: 'projects',
+          // Carga perezosa de la nueva vista de proyectos
+          component: () => import('@/views/project/ProjectsView.vue'),
+        },
+        {
+          path: '/proyectos/:slug',
+          name: 'project-detail',
+          // Carga perezosa de la vista de detalle
+          component: () => import('@/views/project/Slug.vue'),
+        },
+      ],
     },
-    {
-      path: '/about',
-      name: 'about',
-      // route level code-splitting
-      // this generates a separate chunk (About.[hash].js) for this route
-      // which is lazy-loaded when the route is visited.
-      component: () => import('../views/AboutView.vue'),
-    },
+    // Podrías tener otro layout para el panel de administración
+    // {
+    //   path: '/admin',
+    //   component: AdminLayout,
+    //   children: [ ... ]
+    // }
   ],
 })
